refactor: History 파일 구조 재정리(#47) - History 폴더에 History.jsx, History.css 옯김 - hooks 폴더에 useChatHistory 넣었음
