+++ src/App.jsx
@@ -8,7 +8,7 @@ import Redirect from './routes/Redirect';
 import Register from './routes/Register';
 import useReissue from './hooks/useReissue';
 import Chatbot from './routes/Chatbot';
-import History from './routes/History';
+import History from './routes/History/History';
 import ChatStu from './routes/ChatStu';
 import WritingList from './routes/WritingList';
 import WritingSubmit from './routes/WritingSubmit';
